dawn_weverything: Remove some fixable warning suppression flags
-Wundef: Use #if defined(xxx).
-Wundefined-reinterpret-cast: Use std::bit_cast for punning.
-Wfloat-conversion: Use static_cast in PhysicalDeviceVk and tests.
-Wmissing-variable-declarations: Fix Metal and Tint fuzzer code.
-Wshadow-field-in-constructor: Fix shadowing in Core.h.
-Wmissing-noreturn: Add [[noreturn]] to SystemHandle.h.
-Wdeprecated-redundant-constexpr-static-def: Clean up unit tests.
-Wobjc-messaging-id: Cast to NSObject* in NSRef.h.
-Wunique-object-duplication: Fix kNoConfig in DisplayEGL.h.
-Wno-nonportable-system-include-path: add vulkan as system path.

Move -Wno-c++20-extensions to never enable as we are globally
std=c++20 now.

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/320015
Auto-Submit: Shrek Shao <shrekshao@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/dawn/tests/end2end/MultithreadTests.cpp
@@ -1612,7 +1612,8 @@ TEST_P(MultithreadTextureCopyTests, CopyDepthToDepthNoRace) {
 
     std::vector<uint16_t> kExpectedData16(kExpectedData32.size());
     for (size_t i = 0; i < kExpectedData32.size(); ++i) {
-        kExpectedData16[i] = kExpectedData32[i] * std::numeric_limits<uint16_t>::max();
+        kExpectedData16[i] =
+            static_cast<uint16_t>(kExpectedData32[i] * std::numeric_limits<uint16_t>::max());
     }
 
     const size_t kExpectedDataSize16 = kExpectedData16.size() * sizeof(kExpectedData16[0]);
@@ -1680,7 +1681,8 @@ TEST_P(MultithreadTextureCopyTests, CopyBufferToDepthNoRace) {
 
     std::vector<uint16_t> kExpectedData16(kExpectedData32.size());
     for (size_t i = 0; i < kExpectedData32.size(); ++i) {
-        kExpectedData16[i] = kExpectedData32[i] * std::numeric_limits<uint16_t>::max();
+        kExpectedData16[i] =
+            static_cast<uint16_t>(kExpectedData32[i] * std::numeric_limits<uint16_t>::max());
     }
 
     const uint32_t kExpectedDataSize16 = kExpectedData16.size() * sizeof(kExpectedData16[0]);
